moving functions to helpers
